+++ test/videodownlinkbandwidthpolicy/VideoPriorityBasedPolicy.test.ts
@@ -201,6 +201,26 @@ describe('VideoPriorityBasedPolicy', () => {
     });
   });
 
+  describe('default preference', () => {
+    it('use default preference', () => {
+      const policy = new VideoPriorityBasedPolicy(logger);
+      updateIndexFrame(videoStreamIndex, 1, 0, 600);
+      policy.updateIndex(videoStreamIndex);
+      let resub = policy.wantsResubscribe();
+      expect(resub).to.equal(true);
+      let received = policy.chooseSubscriptions();
+      expect(received.array()).to.deep.equal([2]);
+
+      incrementTime(6100);
+      updateIndexFrame(videoStreamIndex, 2, 0, 600);
+      policy.updateIndex(videoStreamIndex);
+      resub = policy.wantsResubscribe();
+      expect(resub).to.equal(true);
+      received = policy.chooseSubscriptions();
+      expect(received.array()).to.deep.equal([2, 4]);
+    });
+  });
+
   describe('chooseRemoteVideoSources', () => {
     it('Can be called if videoPreferences is undefined', () => {
       policy.chooseRemoteVideoSources(VideoPreferences.prepare().build());